requirements.txt
pytest~=8.3.5
pytest-bdd~=8.1.0
playwright~=1.51.0
allure-pytest
pytest-playwright

python-dotenv~=1.1.0
allure-python-commons~=2.14.1